+++ crates/ruma-client-api/src/config/set_global_account_data.rs
@@ -6,8 +6,10 @@ pub mod v3 {
     //! [spec]: https://spec.matrix.org/v1.2/client-server-api/#put_matrixclientv3useruseridaccount_datatype
 
     use ruma_api::ruma_api;
+    use ruma_events::{AnyGlobalAccountDataEventContent, GlobalAccountDataEventContent};
     use ruma_identifiers::UserId;
-    use serde_json::value::RawValue as RawJsonValue;
+    use ruma_serde::Raw;
+    use serde_json::value::to_raw_value as to_raw_json_value;
 
     ruma_api! {
         metadata: {
@@ -38,7 +40,7 @@ pub mod v3 {
             ///
             /// To create a `RawJsonValue`, use `serde_json::value::to_raw_value`.
             #[ruma_api(body)]
-            pub data: &'a RawJsonValue,
+            pub data: Raw<AnyGlobalAccountDataEventContent>,
         }
 
         #[derive(Default)]
@@ -49,7 +51,25 @@ pub mod v3 {
 
     impl<'a> Request<'a> {
         /// Creates a new `Request` with the given data, event type and user ID.
-        pub fn new(data: &'a RawJsonValue, event_type: &'a str, user_id: &'a UserId) -> Self {
+        ///
+        /// # Errors
+        ///
+        /// Since `Request` stores the request body in serialized form, this function can fail if
+        /// `T`s [`Serialize`][serde::Serialize] implementation can fail.
+        pub fn new<T: GlobalAccountDataEventContent>(
+            data: &'a T,
+            event_type: &'a str,
+            user_id: &'a UserId,
+        ) -> serde_json::Result<Self> {
+            Ok(Self { user_id, event_type, data: Raw::from_json(to_raw_json_value(data)?) })
+        }
+
+        /// Creates a new `Request` with the given raw data, event type and user ID.
+        pub fn new_raw(
+            data: Raw<AnyGlobalAccountDataEventContent>,
+            event_type: &'a str,
+            user_id: &'a UserId,
+        ) -> Self {
             Self { user_id, event_type, data }
         }
     }